Rejestracja użytkownika (2 etap). Automatyczne przejście do zmiany hasła po pierwszym logowaniu, zmiana hasła z menu akcji, wylogowanie z listy akcji.

+++ src/main/java/org/sjk/controller/ActionsPageController.java
@@ -1,6 +1,7 @@
 package org.sjk.controller;
 
 import org.sjk.dao.ActionDao;
+import org.sjk.dao.UserDao;
 import org.sjk.dto.Action;
 import org.sjk.dto.User;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,21 +30,55 @@ public class ActionsPageController {
     private ActionDao actionDao;
     @Autowired
     private User currentUser;
+    @Autowired
+    private UserDao userDao;
 
     @RequestMapping(value = "actions",method = RequestMethod.GET)
     public ModelAndView actionsPage(Model model, HttpServletRequest request){
         currentUser=(User)request.getSession().getAttribute("currentUser");
-        List<Action> userActions=actionDao.findAllActions(currentUser.getId());
-        model.addAttribute("actions",userActions);
-        return new ModelAndView("actions_page");
+        if (currentUser!=null){
+            List<Action> userActions=actionDao.findAllActions(currentUser.getId());
+            model.addAttribute("actions",userActions);
+            return new ModelAndView("actions_page");
+        }
+        return new ModelAndView("redirect:/login");
+
     }
     @RequestMapping(value="actionsRemove",method = RequestMethod.POST)
     public ModelAndView removeActions(Model model, HttpServletRequest request){
         currentUser=(User)request.getSession().getAttribute("currentUser");
-        actionDao.removeActionsForUser(currentUser.getId());
-        List<Action> userActions=actionDao.findAllActions(currentUser.getId());
-        model.addAttribute("actions",userActions);
-        return new ModelAndView("actions_page");
+        if (currentUser!=null){
+            actionDao.removeActionsForUser(currentUser.getId());
+            List<Action> userActions=actionDao.findAllActions(currentUser.getId());
+            model.addAttribute("actions",userActions);
+            model.addAttribute("removeActionsSuccess",true);
+            return new ModelAndView("actions_page");
+        }
+        return new ModelAndView("redirect:/login");
+    }
+
+    @RequestMapping(value = "logout",method = RequestMethod.POST)
+    public ModelAndView logoutAction(HttpServletRequest request) throws UnknownHostException {
+        currentUser=(User)request.getSession().getAttribute("currentUser");
+        if (currentUser!=null){
+            String clientIpAddress=request.getHeader("X-FORWARDED-FOR");
+            if (clientIpAddress==null)
+                clientIpAddress=request.getRemoteAddr();
+            if ("127.0.0.1".equals(clientIpAddress)  || "0:0:0:0:0:0:0:1".equals(clientIpAddress) )
+                clientIpAddress= InetAddress.getLocalHost().getHostAddress();
+            userDao.logoutUser(currentUser.getUserName(),clientIpAddress);
+            currentUser=null;
+            request.getSession().setAttribute("currentUser",currentUser);
+        }
+        return new ModelAndView("redirect:/login");
+    }
+
+    @RequestMapping(value="cancel", method = RequestMethod.POST)
+    public ModelAndView cancelAction(HttpServletRequest request){
+        currentUser=(User)request.getSession().getAttribute("currentUser");
+        if (currentUser!=null)
+            return new ModelAndView("redirect:/actions");
+        return new ModelAndView("redirect:/login");
     }
 
 }
